refactor Layout comp into separate pieces and unify spacing
