+++ studio/src/Layout/Settings/index.tsx
@@ -1,0 +1,2 @@
+export { SettingsMenu } from "./SettingsMenu";
+export { SettingsScreen } from "./SettingsScreen";
