+++ studio/src/Layout/SidePanel/index.tsx
@@ -1,0 +1,2 @@
+export { SidePanelTrigger } from "./SidePanelTrigger";
+export { FloatingSidePanel } from "./FloatingSidePanel";
